- Changes based on support routine added to the ABI document, see:   https://github.com/ARM-software/abi-aa/pull/263   - Renamed get_runtime_vl to __arm_get_current_vg   - Also return VG if currently in streaming-mode
- Added static to get_aarch64_cpu_features

- Added aarch64.h and included in sme-abi-init.c
